refactor: quiz 패키지 안에 entry, clear 패키지 생성 후 파일 분리
